chore: new layout replaced with old index
