+++ src/hooks/useVerifiedLocations.tsx
@@ -32,8 +32,8 @@ import useLocation from "./useLocation";
 
 const createGeometry = (res: APIResponseObject): Geometry => ({
   location: {
-    lat: res.loc?.[1] ?? 0,
-    lng: res.loc?.[0] ?? 0,
+    lat: res.loc?.[1] ?? res.lat ?? 0,
+    lng: res.loc?.[0] ?? res.lng ?? 0,
   },
 });
