+++ src/pages/index.tsx
@@ -1,217 +1,148 @@
-import HeadWithMeta from "@/components/base/HeadWithMeta/HeadWithMeta";
-import ClusterPopup from "@/components/UI/ClusterPopup";
-import { Drawer as NewDrawer } from "@/components/UI/Drawer/NewDrawer";
-import Drawer from "@/components/UI/Drawer/Drawer";
-import FilterMenu from "@/components/UI/FilterMenu/FilterMenu";
-import Footer from "@/components/UI/Footer/Footer";
-import FooterBanner from "@/components/UI/Footer/Banner";
-import LocaleSwitch from "@/components/UI/I18n/LocaleSwitch";
-import ScanAreaButton from "@/components/UI/Button/ScanArea";
-import SitesIcon from "@/components/UI/SitesIcon/Icons";
-import {
-  transformers,
-  useVerifiedLocations,
-} from "@/hooks/useVerifiedLocations";
-import { getLocationById } from "@/services/location";
-import { useAreasActions, useShouldFetchNextOption } from "@/stores/areasStore";
-import { MapLayer, useDevice, useMapActions } from "@/stores/mapStore";
-import { useChannelFilterMenuOption, useURLActions } from "@/stores/urlStore";
-import styles from "@/styles/Home.module.css";
-import { APIChannel, APIResponse, ChannelData, DeviceType } from "@/types";
-import { Box } from "@mui/material";
-import Container from "@mui/material/Container";
-import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import dynamic from "next/dynamic";
-import { useRouter } from "next/router";
-import { useEffect } from "react";
-import { parseChannelData } from "@/hooks/useLocation";
+import { Box, SxProps, Theme } from "@mui/material";
+import { HelpViewComponent } from "@/newlayout/components/HelpViewComponent/HelpViewComponent";
+import { serverSideTranslations } from "next-i18next/serverSideTranslations";
+import { ApiClient } from "@/services/ApiClient";
+import { APIResponse, ChannelData } from "@/types";
+import { useState } from "react";
 import { useCopyToClipboard } from "@/hooks/useCopyToClipboard";
+import HeadWithMeta from "@/components/base/HeadWithMeta/HeadWithMeta";
+import { dataFetcher } from "@/services/dataFetcher";
+import { BASE_URL } from "@/utils/constants";
 
-const LeafletMap = dynamic(() => import("@/components/UI/Map"), {
-  ssr: false,
-});
+const MapContent = dynamic(
+  () =>
+    import("@/newlayout/components/map/MapContent").then(
+      (mod) => mod.MapContent
+    ),
+  { ssr: false }
+);
 
-type Props = {
-  deviceType: DeviceType;
-  singleItemDetail: APIResponse;
-  channel: ChannelData;
+const Drawer = dynamic(
+  () => import("@/components/UI/Drawer/NewDrawer").then((mod) => mod.Drawer),
+  { ssr: false }
+);
+// Development overlay container
+const UIElementsOverlay = () => {
+  return (
+    <Box sx={styles.overlay}>
+      <HelpViewComponent />
+    </Box>
+  );
 };
 
-export default function Home({ deviceType, singleItemDetail, channel }: Props) {
-  // gather location data from all channels
-  const channelFilter = useChannelFilterMenuOption();
+interface INHome {
+  reasons: string[];
+  channel: ChannelData;
+  apiResponse: APIResponse;
+}
+
+const NHome = (props: INHome) => {
   const { copyToClipBoard } = useCopyToClipboard();
 
-  const {
-    ahbapLocations,
-    hospitalLocations,
-    foodLocations,
-    teleteyitLocations,
-    satelliteLocations,
-    sahraKitchenLocations,
-    pharmacyLocations,
-    safePlaceLocations,
-    twitterLocations,
-    babalaLocations,
-  } = useVerifiedLocations();
+  const [reasons] = useState(() => props.reasons);
+  const [locations, setLocations] = useState<ChannelData[]>(() => []);
 
-  const router = useRouter();
-
-  const onLanguageChange = (newLocale: string) => {
-    const { pathname, asPath, query } = router;
-    router.push({ pathname, query }, asPath, { locale: newLocale });
-  };
-
-  // adds timestamp param to the URL query if data is from babala
-  const { setTimeStamp } = useURLActions();
-
-  // FIXME: device handling, kinda weird how this is handled right now...
-  const device = useDevice();
-  const isMobile = device === "mobile";
-  const { setDevice } = useMapActions();
-  useEffect(() => {
-    setDevice(deviceType);
-  }, [deviceType, setDevice]);
-
-  // supposed to handle re-fetching after filter selection (FIXME: FilterTimeMenu.tsx)
-  const { setShouldFetchNextOption } = useAreasActions();
-  const resetShouldFetchNextOption = () => setShouldFetchNextOption(false);
-  const shouldFetchNextOption = useShouldFetchNextOption();
-
-  // global error handling from location data fetching
   return (
     <>
-      <HeadWithMeta singleItemDetail={singleItemDetail} />
-      <main className={styles.main} id="prod-layout">
-        <Container maxWidth={false} disableGutters>
-          <div
-            style={{
-              position: "fixed",
-              right: "10px",
-              top: "15px",
-              zIndex: "502",
-            }}
-          >
-            <div
-              style={{
-                display: "flex",
-                flexDirection: "column",
-                alignItems: "flex-end",
-                gap: 2,
-              }}
-            >
-              <FilterMenu>
-                <FilterMenu.Channel />
-                <FilterMenu.Time
-                  onChangeTime={setTimeStamp}
-                  shouldFetchNextOption={shouldFetchNextOption}
-                  resetShouldFetchNextOption={resetShouldFetchNextOption}
-                />
-                <FilterMenu.Reason />
-              </FilterMenu>
-            </div>
-          </div>
-          <LeafletMap
-            locations={{
-              [MapLayer.Ahbap]: ahbapLocations,
-              [MapLayer.Food]: foodLocations,
-              [MapLayer.Hospital]: hospitalLocations,
-              [MapLayer.Satellite]: satelliteLocations,
-              [MapLayer.SahraMutfak]: sahraKitchenLocations,
-              [MapLayer.Pharmacy]: pharmacyLocations,
-              [MapLayer.SafePlaces]: safePlaceLocations,
-              [MapLayer.Teleteyit]: teleteyitLocations,
-              [MapLayer.Markers]:
-                channelFilter === "twitter"
-                  ? twitterLocations
-                  : channelFilter === "babala"
-                  ? babalaLocations
-                  : twitterLocations.concat(babalaLocations),
-            }}
-          />
-          {/* FIXME: move it to a component */}
-          <Box
-            sx={{
-              display: "flex",
-              padding: "0",
-              borderRadius: "10px",
-              position: "absolute",
-              bottom: isMobile ? "30px" : "90px",
-              right: isMobile ? "10px" : "26px",
-              zIndex: 500,
-            }}
-          >
-            <LocaleSwitch
-              current={router.locale || "tr"}
-              onChange={onLanguageChange}
-              mobile={isMobile}
-            />
-          </Box>
-          {!isMobile && <SitesIcon></SitesIcon>}
-          {/* FIXME: Move it to a component */}
-          <Box
-            sx={{
-              position: "fixed",
-              top: { md: "15px" },
-              bottom: { xs: "88px", md: "unset" },
-              left: "50%",
-              marginLeft: "-105px",
-              zIndex: "502",
-              display: "flex",
-              flexDirection: "column",
-              rowGap: "8px",
-              width: "210px",
-            }}
-          >
-            <ScanAreaButton />
-          </Box>
-        </Container>
-        {channel && (
-          <NewDrawer
-            data={channel}
-            onCopyBillboard={(_clipped) => copyToClipBoard(_clipped as string)}
-          />
-        )}
-
-        <Drawer />
-        <ClusterPopup />
-        <FooterBanner />
-        <Footer />
+      <HeadWithMeta singleItemDetail={props.apiResponse} />
+      <main id="new-layout">
+        <UIElementsOverlay />
+        <MapContent
+          reasons={reasons}
+          locations={locations}
+          setLocations={setLocations}
+        />
+        <Drawer
+          data={props.channel}
+          onCopyBillboard={(_clipped) => copyToClipBoard(_clipped as string)}
+        />
       </main>
     </>
   );
-}
+};
+export default NHome;
 
 export async function getServerSideProps(context: any) {
+  const client = new ApiClient({ url: "https://apigo.afetharita.com" });
+  const searchParams = new URLSearchParams(context.query);
+  let redirect = false;
+
+  const reasons = await client.fetchReasons();
+
+  let channel: ChannelData | null = null;
+  let apiResponse: APIResponse | null = null;
+  if (context.query.id) {
+    apiResponse = (await dataFetcher(
+      `${BASE_URL}/feeds/${context.query.id}`
+    )) as APIResponse;
+    channel = await client.fetchLocationByID(context.query.id);
+    if (!channel) {
+      searchParams.delete("id");
+      redirect = true;
+    }
+  }
+
+  // Pass data to the page via props
   const UA = context.req.headers["user-agent"];
   const isMobile = Boolean(
     UA.match(
       /Android|BlackBerry|iPhone|iPad|iPod|Opera Mini|IEMobile|WPDesktop/i
     )
   );
-
-  let itemDetail: APIResponse | null = null;
-  if (context.query.id) {
-    itemDetail = await getLocationById(context.query.id);
+  if (context.query.reasons === "" || !context.query.reasons) {
+    searchParams.delete("reasons");
+    searchParams.append("reasons", reasons.join(","));
+    redirect = true;
   }
 
-  let channel: ChannelData | undefined = undefined;
-  if (itemDetail) {
-    channel = parseChannelData(itemDetail, {
-      transformResponse:
-        transformers[itemDetail.channel.toLowerCase() as APIChannel],
-    });
+  if (redirect) {
+    return {
+      redirect: {
+        permanent: false,
+        destination: "/?" + searchParams.toString(),
+      },
+    };
   }
-
   return {
     props: {
       ...(await serverSideTranslations(context.locale, ["common", "home"])),
       deviceType: isMobile ? "mobile" : "desktop",
-      singleItemDetail: {
-        ...itemDetail,
-        ...context.query,
-      },
-      channel: channel ?? null,
+      reasons,
+      channel,
+      apiResponse,
     },
   };
 }
+
+interface IStyles {
+  [key: string]: SxProps<Theme>;
+}
+
+const styles: IStyles = {
+  overlay: (theme: Theme) => ({
+    display: "flex",
+    position: "fixed",
+    flexDirection: "column",
+    top: "85px",
+    left: "55px",
+    zIndex: 1100,
+    pointerEvents: "none",
+    [theme.breakpoints.down("sm")]: {
+      top: "0px",
+      left: "0px",
+    },
+  }),
+  fixedMidBottom: () => ({
+    position: "fixed",
+    bottom: "0px",
+    left: "0px",
+    width: "100%",
+    height: "110px",
+    zIndex: 1030,
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    pointerEvents: "none",
+  }),
+};
